agents/supplier-agent/requirements.txt
a2a-sdk[http-server]
psycopg2-binary
pgvector
uvicorn
google-cloud-aiplatform
python-dotenv
